% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{bconf}
\alias{bconf}
\alias{print.bconf}
\title{Print current configuations}
\format{
An object of class \code{bconf} of length 1.
}
\usage{
bconf

\method{print}{bconf}(x, ...)
}
\arguments{
\item{x}{A `bconf` object.}

\item{...}{Other parameters.}
}
\value{
A `bconf` object.
}
\description{
Print current configuations
}
\details{
This function is only for printing. Use [`bsub_opt()`] to change configurations.

You simply type ``bconf`` (without the brackets) in the interactive R console.
}
\examples{
bconf
}
\keyword{datasets}
