\name{bjobs_done}
\alias{bjobs_done}
\title{
Finished jobs
}
\description{
Finished jobs
}
\usage{
bjobs_done(max = Inf, filter = NULL)
}
\arguments{

  \item{max}{Maximal number of jobs.}
  \item{filter}{Regular expression to filter on job names.}

}
\details{
You can directly type \code{bjobs_done} without parentheses which runs \code{\link{bjobs_done}} with defaults.
}
\value{
The same output format as \code{\link{bjobs}}.
}
\examples{
# There is no example
NULL

}
