\name{bjobs_pending}
\alias{bjobs_pending}
\title{
Pending jobs
}
\description{
Pending jobs
}
\usage{
bjobs_pending(max = Inf, filter = NULL)
}
\arguments{

  \item{max}{Maximal number of jobs}
  \item{filter}{Regular expression to filter on job names}

}
\details{
You can directly type \code{bjobs_pending} without parentheses which runs \code{\link{bjobs_pending}} with defaults.
}
\examples{
# There is no example
NULL

}
