\name{bjobs_running}
\alias{bjobs_running}
\title{
Running jobs
}
\description{
Running jobs
}
\usage{
bjobs_running(max = Inf, filter = NULL)
}
\arguments{

  \item{max}{Maximal number of jobs.}
  \item{filter}{Regular expression to filter on job names.}

}
\details{
You can directly type \code{bjobs_running} without parentheses which runs \code{\link{bjobs_running}} with defaults.
}
\value{
The same output format as \code{\link{bjobs}}.
}
\examples{
# There is no example
NULL

}
