% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cmd.R
\name{bkill}
\alias{bkill}
\title{Kill jobs}
\usage{
bkill(job_id, filter = NULL)
}
\arguments{
\item{job_id}{A vector of job IDs or a data frame returned by [`bjobs()`].}

\item{filter}{Regular expression on job names (only the running and pending jobs). It is
only used when `job_id` is not set.}
}
\description{
Kill jobs
}
\examples{
\dontrun{
job_id = c(10000000, 10000001, 10000002)  # job ids can be get from `bjobs()`
bkill(job_id)
# kill all jobs (running and pending) of which the names contain "example"
bkill(filter = "example") 
}
}
