\name{brecent}
\alias{brecent}
\title{
Recent jobs with status "all"
}
\description{
Recent jobs with status "all"
}
\usage{
brecent(max = 20, filter = NULL)
}
\arguments{

  \item{max}{Maximal number of recent jobs.}
  \item{filter}{Regular expression to filter on job names.}

}
\details{
You can directly type \code{brecent} without parentheses which runs \code{\link{brecent}} with defaults.
}
\value{
The same output format as \code{\link{bjobs}}.
}
\examples{
# There is no example
NULL

}
