\name{dep_dot}
\alias{dep_dot}
\title{
The DOT code of a job dependency graph
}
\description{
The DOT code of a job dependency graph
}
\usage{
dep_dot(job_id, job_tb = NULL)
}
\arguments{

  \item{job_id}{A job id.}
  \item{job_tb}{A table from \code{\link{bjobs}}. Optional.}

}
\examples{
# There is no example
NULL
}
