% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{formatFileSize}
\alias{formatFileSize}
\title{Format file size}
\usage{
formatFileSize(table, columns)
}
\arguments{
\item{table}{Internally used.}

\item{columns}{Internally used.}
}
\description{
Format file size
}
