% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{job_status_by_name}
\alias{job_status_by_name}
\alias{job_status_by_id}
\title{Job status by ID or name}
\usage{
job_status_by_name(job_name)

job_status_by_id(job_id)
}
\arguments{
\item{job_name}{A single job name.}

\item{job_id}{A single job ID.}
}
\description{
Job status by ID or name
}
\examples{
\dontrun{
job_status_by_name("example")
job_status_by_id(123456)
}
}
