\name{job_status_by_id}
\alias{job_status_by_id}
\title{
Job status by id
}
\description{
Job status by id
}
\usage{
job_status_by_id(job_id)
}
\arguments{

  \item{job_id}{The job id}

}
\value{
If the job has been deleted from the database, it returns MISSING.
}
\examples{
# There is no example
NULL
}
