% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{list_dump_files}
\alias{list_dump_files}
\title{Check whether there are dump files}
\usage{
list_dump_files(print = TRUE)
}
\arguments{
\item{print}{Whether to print messages.}
}
\value{
A vector of file names.
}
\description{
Check whether there are dump files
}
\details{
For the failed jobs, LSF cluster might generate a core dump file and R might generate a \code{.RDataTmp} file.

Note if you manually set working directory in your R code/script, the R dump file can be not caught.
}
\examples{
\dontrun{
list_dump_files()
}
}
