% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{monitor}
\alias{monitor}
\title{Browser-based interactive job monitor}
\usage{
monitor()
}
\description{
Browser-based interactive job monitor
}
\details{
The monitor is implemented as a shiny app.
}
\examples{
\dontrun{
monitor()
}
}
