\name{plot_dependency}
\alias{plot_dependency}
\title{
Plot the job dependency tree
}
\description{
Plot the job dependency tree
}
\usage{
plot_dependency(job_id, job_tb = NULL)
}
\arguments{

  \item{job_id}{A job ID.}
  \item{job_tb}{A table from \code{\link{bjobs}}. Optional.}

}
\value{
No value is returned.
}
\examples{
# There is no example
NULL

}
