\name{print.bconf}
\alias{print.bconf}
\title{
Print the configurations
}
\description{
Print the configurations
}
\usage{
\method{print}{bconf}(x, ...)
}
\arguments{

  \item{x}{A bconf object}
  \item{...}{Other parameters}

}
\value{
No value is returned.
}
\examples{
# There is no example
NULL
}
