% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsub.R
\name{random_job}
\alias{random_job}
\title{Submit a random job}
\usage{
random_job(name, ...)
}
\arguments{
\item{name}{Job name.}

\item{...}{Pass to [`bsub_chunk()`].}
}
\value{
A job ID.
}
\description{
Submit a random job
}
\details{
It simply runs `Sys.sleep(30)` in the job.
}
\examples{
\dontrun{
random_job()
}
}
