% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/status.R
\name{retrieve_var}
\alias{retrieve_var}
\title{Retrieve saved variable}
\usage{
retrieve_var(job_id, job_name = NULL, wait = 30)
}
\arguments{
\item{job_id}{A single job ID.}

\item{job_name}{A single Job name. Since jobs may be the same names, it is more suggested to set `job_id`.}

\item{wait}{Seconds to wait.}
}
\value{
The retrieved object.
}
\description{
Retrieve saved variable
}
\details{
It retrieve the saved variable in `bsub_chunk()` when ``save_rds = TRUE`` is set.
}
\examples{
\dontrun{
job_id = bsub_chunk(name = "example", save_var = TRUE,
{
    Sys.sleep(10)
    1+1
})
retrieve_var(job_id)
}
}
