% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cmd.R
\name{run_cmd}
\alias{run_cmd}
\title{Run command on submission node}
\usage{
run_cmd(cmd, print = FALSE)
}
\arguments{
\item{cmd}{A single-line command.}

\item{print}{Whether to print output from the command.}
}
\value{
The output of the command.
}
\description{
Run command on submission node
}
\details{
If current node is not the submission node, the command is executed via ssh.
}
\examples{
\dontrun{
# run pwd on remote node
run_cmd("pwd")
}
}
