% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssh.R
\name{ssh_connect}
\alias{ssh_connect}
\alias{ssh_disconnect}
\title{Connect to submisstion via ssh}
\usage{
ssh_connect(...)

ssh_disconnect()
}
\arguments{
\item{...}{Pass to \code{\link[ssh:ssh]{ssh::ssh_connect()}}.}
}
\description{
Connect to submisstion via ssh
}
\examples{
# ssh is automatically connected. To manually connect ssh, run:
\dontrun{
ssh_connect()
ssh_disconnect()
}
# where the user name is the one you set in `bsub_opt$user` and
# the node is the one you set in `bsub_opt$login_node`.
}
