\name{ssh_disconnect}
\alias{ssh_disconnect}
\title{
Disconnect ssh connection
}
\description{
Disconnect ssh connection
}
\usage{
ssh_disconnect()
}
\value{
No value is returned.
}
\examples{
# Normally you don't need to manually run this function. The ssh is automatically
# disconnected when the package is detached.
# To manually disconnect ssh, run:
\dontrun{
ssh_disconnect()
}
}
