% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monitor.R
\name{bjobs_barplot}
\alias{bjobs_barplot}
\alias{bjobs_timeline}
\title{Visualize statistics of jobs}
\usage{
bjobs_barplot(
  status = c("RUN", "EXIT", "PEND", "DONE"),
  filter = NULL,
  job_tb = NULL
)

bjobs_timeline(
  status = c("RUN", "EXIT", "PEND", "DONE"),
  filter = NULL,
  job_tb = NULL
)
}
\arguments{
\item{status}{Status of the jobs. Use "all" for all jobs.}

\item{filter}{Regular expression to filter on job names.}

\item{job_tb}{A data frame returned by [`bjobs()`]. Only internally used.}
}
\description{
Visualize statistics of jobs
}
\details{
`bjobs_barplot()` draws barplots of number of jobs per day.
`bjobs_timeline()` draws segments of duration of jobs. In the plot, each segment represents
a job and the width of the segment correspond to its duration.
}
