\name{wait_jobs}
\alias{wait_jobs}
\title{
Wait until all jobs are finished
}
\description{
Wait until all jobs are finished
}
\usage{
wait_jobs(job_name, output_dir = bsub_opt$output_dir, wait = 30)
}
\arguments{

  \item{job_name}{A vector of job names.}
  \item{output_dir}{Output dir.}
  \item{wait}{Seconds to wait.}

}
\value{
No value is returned.
}
\examples{
# There is no example
NULL

}
