% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global.R
\docType{data}
\name{bconf}
\alias{bconf}
\alias{print.bconf}
\title{Print current configurations}
\format{
An object of class \code{bconf} of length 1.
}
\usage{
bconf

\method{print}{bconf}(x, ...)
}
\arguments{
\item{x}{A \code{bconf} object.}

\item{...}{Other parameters.}
}
\value{
A \code{bconf} object.
}
\description{
Print current configurations
}
\details{
This function is only for printing. Use \code{\link[=bsub_opt]{bsub_opt()}} to change configurations.

You simply type \code{bconf} (without the brackets) in the interactive R console.
}
\examples{
bconf
}
\keyword{datasets}
