\name{bconf}
\docType{data}
\alias{bconf}
\title{
Print current configuation
}
\description{
Print current configuation
}
\usage{
bconf
}
\details{
This function is only for printing. Use \code{\link{bsub_opt}} to change configurations.

You simply type \code{bconf} (without the brackets) in the interactive R console.
}
\value{
A \code{bconf} object.
}
\examples{
bconf
}
