\name{brecent}
\alias{brecent}
\title{
Recent jobs
}
\description{
Recent jobs
}
\usage{
brecent(max = 20, filter = NULL)
}
\arguments{

  \item{max}{Maximal number of recent jobs}
  \item{filter}{Regular expression to filter on job names}

}
\details{
It is \code{bjobs(status = "all", max = 20, filter = ...)}.
}
\examples{
# There is no example
NULL
}
