% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_bsub}
\alias{config_bsub}
\title{Configure bsub global options}
\usage{
config_bsub()
}
\description{
Configure bsub global options
}
\details{
It sets the submission nodes, user name and how to call `Rscript`.
}
