\name{job_log}
\alias{job_log}
\title{
Log for the running/finished/failed job
}
\description{
Log for the running/finished/failed job
}
\usage{
job_log(job_id, print = TRUE)
}
\arguments{

  \item{job_id}{The job id}
  \item{print}{Whether print the log message.}

}
\value{
The log message as a vector.
}
\examples{
# There is no example
NULL

}
