\name{monitor}
\alias{monitor}
\title{
A browser-based interactive job monitor
}
\description{
A browser-based interactive job monitor
}
\usage{
monitor()
}
\details{
The monitor is implemented as a shiny app.
}
\value{
No value is returned.
}
\examples{
\dontrun{
# simply run:
monitor
# or
monitor()
}
}
