% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsub.R
\name{random_job}
\alias{random_job}
\title{Submit a random job}
\usage{
random_job(name, secs = 30, ...)
}
\arguments{
\item{name}{Job name.}

\item{secs}{Seconds to sleep.}

\item{...}{Pass to \code{\link[=bsub_chunk]{bsub_chunk()}}.}
}
\value{
A job ID.
}
\description{
Submit a random job
}
\details{
It simply runs \code{Sys.sleep(secs)} in the job.
}
\examples{
\dontrun{
random_job()
}
}
