% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsub.R
\name{random_job}
\alias{random_job}
\title{Submit a random job}
\usage{
random_job(name, sleep = 30, ...)
}
\arguments{
\item{name}{Job name.}

\item{sleep}{The random job simply runs [`Sys.sleep()`]. This argument controls the number of seconds to sleep.}

\item{...}{Pass to [`bsub_chunk()`].}
}
\value{
A job ID.
}
\description{
Submit a random job
}
\examples{
\dontrun{
random_job()
random_job(name = "test", sleep = 600) # 10 min
}
}
