% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun.R
\name{job_rerun}
\alias{job_rerun}
\alias{pipeline_rerun}
\title{Rerun a job}
\usage{
job_rerun(job_id, dependency = character(0), verbose = TRUE, job_tb = NULL)

pipeline_rerun(job_id, skip_done = TRUE, verbose = TRUE)
}
\arguments{
\item{job_id}{A single job ID. In `pipeline_rerun()`, it is the job ID of any one job in the pipeline.}

\item{dependency}{A vector of job IDs that current job depends on.}

\item{verbose}{Whether to print messages.}

\item{job_tb}{The data frame returned from [`bjobs()`], internally used.}

\item{skip_done}{Whether to skip done jobs.}
}
\description{
Rerun a job
}
